// src/models/estabelecimento.ts
export class Estabelecimento {
  id: number;
  atividade: string;
  razao: string;
  fantasia: string;
  cnpj: string;
  insc: string;
  ieourg: string;
  endereco: string;
  bairro: string;
  cep: string;
  telefone: string;
  celular: string;
  obs: string;
  data_inspecao: string;
  data_retorno: string;
  status: string;
  data_licenca: string;
  pessoa: string;
  pref: string;
  licenca: string;
  obs_licenca: string;
  cnae: string;
  email: string;
}